chore: cleanup structure to less nested folder
