--- hdp/src/provider/envelope/evm/datalake/mod.rs
@@ -1,28 +1,0 @@
-use std::collections::HashSet;
-
-use crate::primitives::processed_types::{
-    account::ProcessedAccount, header::ProcessedHeader, mmr::MMRMeta, receipt::ProcessedReceipt,
-    storage::ProcessedStorage, transaction::ProcessedTransaction,
-};
-
-use alloy::primitives::U256;
-
-pub mod block_sampled;
-pub mod transactions;
-
-pub struct FetchedDatalake {
-    /// Targeted datalake's compiled results
-    pub values: Vec<U256>,
-    /// Headers related to the datalake
-    pub headers: HashSet<ProcessedHeader>,
-    /// Accounts related to the datalake
-    pub accounts: HashSet<ProcessedAccount>,
-    /// Storages related to the datalake
-    pub storages: HashSet<ProcessedStorage>,
-    /// Transactions related to the datalake
-    pub transactions: HashSet<ProcessedTransaction>,
-    /// Transaction receipts related to the datalake
-    pub transaction_receipts: HashSet<ProcessedReceipt>,
-    /// MMR meta data related to the headers
-    pub mmr_metas: HashSet<MMRMeta>,
-}
