+++ hdp/src/provider/error.rs
