+++ hdp/src/provider/key.rs
@@ -3,13 +3,103 @@
 //!
 //! TODO: need to sync with how bootloader will emit the keys
 
-use std::{hash::Hash, str::FromStr};
+use std::{
+    collections::{HashMap, HashSet},
+    hash::Hash,
+    str::FromStr,
+};
 
 use crate::primitives::task::datalake::{
     block_sampled::BlockSampledCollectionType, transactions::TransactionsCollectionType,
 };
 use alloy::primitives::{Address, BlockNumber, ChainId, Keccak256, StorageKey, B256};
 use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Default)]
+/// This is keys that are categorized into different subsets of keys.
+pub struct CategorizedFetchKeys {
+    pub headers: HashSet<HeaderMemorizerKey>,
+    pub accounts: HashSet<AccountMemorizerKey>,
+    pub storage: HashSet<StorageMemorizerKey>,
+    pub txs: HashSet<TxMemorizerKey>,
+    pub tx_receipts: HashSet<TxReceiptMemorizerKey>,
+}
+
+impl CategorizedFetchKeys {
+    pub fn new(
+        headers: HashSet<HeaderMemorizerKey>,
+        accounts: HashSet<AccountMemorizerKey>,
+        storage: HashSet<StorageMemorizerKey>,
+        txs: HashSet<TxMemorizerKey>,
+        tx_receipts: HashSet<TxReceiptMemorizerKey>,
+    ) -> Self {
+        Self {
+            headers,
+            accounts,
+            storage,
+            txs,
+            tx_receipts,
+        }
+    }
+}
+
+/// Categorize fetch keys
+/// This is require to initiate multiple provider for different chain and fetch keys types
+pub fn categorize_fetch_keys(
+    fetch_keys: Vec<FetchKeyEnvelope>,
+) -> Vec<(ChainId, CategorizedFetchKeys)> {
+    let mut chain_id_map: HashMap<u64, CategorizedFetchKeys> = std::collections::HashMap::new();
+
+    for key in fetch_keys {
+        let chain_id = key.get_chain_id();
+        let target_categorized_fetch_keys = chain_id_map.entry(chain_id).or_default();
+
+        match key {
+            FetchKeyEnvelope::Header(header_key) => {
+                target_categorized_fetch_keys.headers.insert(header_key);
+            }
+            FetchKeyEnvelope::Account(account_key) => {
+                target_categorized_fetch_keys
+                    .headers
+                    .insert(HeaderMemorizerKey::new(
+                        account_key.chain_id,
+                        account_key.block_number,
+                    ));
+                target_categorized_fetch_keys.accounts.insert(account_key);
+            }
+            FetchKeyEnvelope::Storage(storage_key) => {
+                target_categorized_fetch_keys
+                    .headers
+                    .insert(HeaderMemorizerKey::new(
+                        storage_key.chain_id,
+                        storage_key.block_number,
+                    ));
+                target_categorized_fetch_keys.storage.insert(storage_key);
+            }
+            FetchKeyEnvelope::Tx(tx_key) => {
+                target_categorized_fetch_keys
+                    .headers
+                    .insert(HeaderMemorizerKey::new(
+                        tx_key.chain_id,
+                        tx_key.block_number,
+                    ));
+                target_categorized_fetch_keys.txs.insert(tx_key);
+            }
+            FetchKeyEnvelope::TxReceipt(tx_receipt_key) => {
+                target_categorized_fetch_keys
+                    .headers
+                    .insert(HeaderMemorizerKey::new(
+                        tx_receipt_key.chain_id,
+                        tx_receipt_key.block_number,
+                    ));
+                target_categorized_fetch_keys
+                    .tx_receipts
+                    .insert(tx_receipt_key);
+            }
+        }
+    }
+    chain_id_map.into_iter().collect()
+}
 
 /// Key for fetching block header from provider.
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Hash)]
