+++ hdp/src/provider/types.rs
@@ -3,6 +3,14 @@
 //!
 //! We need this type to bind encoded transaction and receipts to the block number and proofs.
 
+use std::collections::HashSet;
+
+use crate::primitives::processed_types::{
+    account::ProcessedAccount, header::ProcessedHeader, mmr::MMRMeta, receipt::ProcessedReceipt,
+    storage::ProcessedStorage, transaction::ProcessedTransaction,
+};
+
+use alloy::primitives::U256;
 use alloy::{
     consensus::TxType,
     primitives::{BlockNumber, Bytes, TxIndex},
@@ -60,4 +68,21 @@ impl FetchedTransactionReceiptProof {
             tx_type,
         }
     }
+}
+
+pub struct FetchedDatalake {
+    /// Targeted datalake's compiled results
+    pub values: Vec<U256>,
+    /// Headers related to the datalake
+    pub headers: HashSet<ProcessedHeader>,
+    /// Accounts related to the datalake
+    pub accounts: HashSet<ProcessedAccount>,
+    /// Storages related to the datalake
+    pub storages: HashSet<ProcessedStorage>,
+    /// Transactions related to the datalake
+    pub transactions: HashSet<ProcessedTransaction>,
+    /// Transaction receipts related to the datalake
+    pub transaction_receipts: HashSet<ProcessedReceipt>,
+    /// MMR meta data related to the headers
+    pub mmr_metas: HashSet<MMRMeta>,
 }
